Complete A-Assertions
There are several points in the code where we assume certain variables
to be holding values satisfying certain assumptions. However, these
assumptions are not being checked for explicitly.

This can be dangerous, as if these assumptions do not hold, then the
code may not behave as expected.

Let's insert assertions at these critical points.

This will allow us to explicitly check that these assumptions are being
satisfied, preventing unexpected code behaviour.

This also completes the A-Assertions increment.