Complete A-CodeQuality
There are some code quality violations identified.

This needs to be changed to allow for better readability.

I have broken down a long function body (handleCommand) into smaller
functions for greater modularity.

This is to prevent excessive nesting, and abstract out the
implementation details in one more layer.

This also completes the A-CodeQuality task.